+++ Source/JavaScriptCore/runtime/JSBigInt.h
@@ -442,16 +442,7 @@ public:
 #endif
 
     static JSValue toNumberHeap(JSBigInt*);
-    static JSValue toNumber(JSValue bigInt)
-    {
-        ASSERT(bigInt.isBigInt());
-#if USE(BIGINT32)
-        if (bigInt.isBigInt32())
-            return jsNumber(bigInt.bigInt32AsInt32());
-#endif
-        return toNumberHeap(jsCast<JSBigInt*>(bigInt));
-    }
-
+    inline static JSValue toNumber(JSValue); // Defined in JSBigIntInlines.h
 
     static JSValue asIntN(JSGlobalObject*, uint64_t numberOfBits, JSBigInt*);
     static JSValue asUintN(JSGlobalObject*, uint64_t numberOfBits, JSBigInt*);
@@ -460,25 +451,8 @@ public:
     static JSValue asUintN(JSGlobalObject*, uint64_t numberOfBits, int32_t bigIntAsInt32);
 #endif
 
-    static uint64_t toBigUInt64(JSValue bigInt)
-    {
-        ASSERT(bigInt.isBigInt());
-#if USE(BIGINT32)
-        if (bigInt.isBigInt32())
-            return static_cast<uint64_t>(static_cast<int64_t>(bigInt.bigInt32AsInt32()));
-#endif
-        return toBigUInt64Heap(bigInt.asHeapBigInt());
-    }
-
-    static int64_t toBigInt64(JSValue bigInt)
-    {
-        ASSERT(bigInt.isBigInt());
-#if USE(BIGINT32)
-        if (bigInt.isBigInt32())
-            return static_cast<int64_t>(bigInt.bigInt32AsInt32());
-#endif
-        return static_cast<int64_t>(toBigUInt64Heap(bigInt.asHeapBigInt()));
-    }
+    inline static uint64_t toBigUInt64(JSValue); // Defined in JSBigIntInlines.h
+    inline static int64_t toBigInt64(JSValue); // Defined in JSBigIntInlines.h
 
     Digit digit(unsigned);
     void setDigit(unsigned, Digit); // Use only when initializing.
@@ -493,7 +467,7 @@ public:
         return hashSlow();
     }
 
-    static std::optional<double> tryExtractDouble(JSValue);
+    inline static std::optional<double> tryExtractDouble(JSValue); // Defined in JSBigIntInlines.h
 
     inline bool isZero() const
     {
@@ -700,44 +674,6 @@ ALWAYS_INLINE JSValue tryConvertToBigInt32(JSBigInt* bigInt)
 #endif
 
     return bigInt;
-}
-
-ALWAYS_INLINE std::optional<double> JSBigInt::tryExtractDouble(JSValue value)
-{
-    if (value.isNumber())
-        return value.asNumber();
-
-    if (!value.isBigInt())
-        return std::nullopt;
-
-#if USE(BIGINT32)
-    if (value.isBigInt32())
-        return value.bigInt32AsInt32();
-#endif
-
-    ASSERT(value.isHeapBigInt());
-    JSBigInt* bigInt = value.asHeapBigInt();
-    if (!bigInt->length())
-        return 0;
-
-    uint64_t integer = 0;
-    if constexpr (sizeof(Digit) == 8) {
-        if (bigInt->length() != 1)
-            return std::nullopt;
-        integer = bigInt->digit(0);
-    } else {
-        ASSERT(sizeof(Digit) == 4);
-        if (bigInt->length() > 2)
-            return std::nullopt;
-        integer = bigInt->digit(0);
-        if (bigInt->length() == 2)
-            integer |= (static_cast<uint64_t>(bigInt->digit(1)) << 32);
-    }
-
-    if (integer <= maxSafeIntegerAsUInt64())
-        return (bigInt->sign()) ? -static_cast<double>(integer) : static_cast<double>(integer);
-
-    return std::nullopt;
 }
 
 } // namespace JSC
