+++ Source/JavaScriptCore/runtime/JSCJSValueInlines.h
@@ -36,6 +36,7 @@ WTF_ALLOW_UNSAFE_BUFFER_USAGE_BEGIN
 #include <JavaScriptCore/InternalFunction.h>
 #include <JavaScriptCore/JSBigInt.h>
 #include <JavaScriptCore/JSCJSValue.h>
+#include <JavaScriptCore/JSCJSValueCellInlines.h>
 #include <JavaScriptCore/JSCellInlines.h>
 #include <JavaScriptCore/JSFunction.h>
 #include <JavaScriptCore/JSGlobalProxy.h>
@@ -48,6 +49,27 @@ WTF_ALLOW_UNSAFE_BUFFER_USAGE_BEGIN
 WTF_ALLOW_UNSAFE_BUFFER_USAGE_END
 
 namespace JSC {
+
+inline JSValue jsNumber(double d)
+{
+    ASSERT(JSValue(d).isNumber());
+    ASSERT(!isImpureNaN(d));
+    return JSValue(d);
+}
+
+inline JSValue jsNumber(const MediaTime& t)
+{
+    return jsNumber(t.toDouble());
+}
+
+inline JSValue::JSValue(double d)
+{
+    if (canBeStrictInt32(d)) {
+        *this = JSValue(static_cast<int32_t>(d));
+        return;
+    }
+    *this = JSValue(EncodeAsDouble, d);
+}
 
 ALWAYS_INLINE int32_t JSValue::toInt32(JSGlobalObject* globalObject) const
 {
@@ -95,40 +117,6 @@ inline uint64_t JSValue::toIndex(JSGlobalObject* globalObject, ASCIILiteral erro
     RELEASE_AND_RETURN(scope, d);
 }
 
-// https://tc39.es/proposal-temporal/#sec-tointegerwithtruncation
-inline double JSValue::toIntegerWithTruncation(JSGlobalObject* globalObject) const
-{
-    if (isInt32())
-        return asInt32();
-    return trunc(toNumber(globalObject) + 0.0);
-}
-
-// https://tc39.es/ecma262/#sec-tointegerorinfinity
-inline double JSValue::toIntegerOrInfinity(JSGlobalObject* globalObject) const
-{
-    if (isInt32())
-        return asInt32();
-    double d = toNumber(globalObject);
-    return isnan(d) ? 0.0 : trunc(d) + 0.0;
-}
-
-inline bool JSValue::isUInt32() const
-{
-    return isInt32() && asInt32() >= 0;
-}
-
-inline uint32_t JSValue::asUInt32() const
-{
-    ASSERT(isUInt32());
-    return asInt32();
-}
-
-inline double JSValue::asNumber() const
-{
-    ASSERT(isNumber());
-    return isInt32() ? asInt32() : asDouble();
-}
-
 inline std::optional<uint32_t> JSValue::tryGetAsUint32Index()
 {
     if (isUInt32()) {
@@ -157,564 +145,28 @@ inline std::optional<int32_t> JSValue::tryGetAsInt32()
     return std::nullopt;
 }
 
-inline JSValue jsNaN()
-{
-    return JSValue(JSValue::EncodeAsDouble, PNaN);
-}
-
-inline JSValue::JSValue(char i)
-{
-    *this = JSValue(static_cast<int32_t>(i));
-}
-
-inline JSValue::JSValue(unsigned char i)
-{
-    *this = JSValue(static_cast<int32_t>(i));
-}
-
-inline JSValue::JSValue(short i)
-{
-    *this = JSValue(static_cast<int32_t>(i));
-}
-
-inline JSValue::JSValue(unsigned short i)
-{
-    *this = JSValue(static_cast<int32_t>(i));
-}
-
-inline JSValue::JSValue(unsigned i)
-{
-    if (static_cast<int32_t>(i) < 0) {
-        *this = JSValue(EncodeAsDouble, static_cast<double>(i));
-        return;
-    }
-    *this = JSValue(static_cast<int32_t>(i));
-}
-
-inline JSValue::JSValue(long i)
-{
-    if (static_cast<int32_t>(i) != i) {
-        *this = JSValue(EncodeAsDouble, static_cast<double>(i));
-        return;
-    }
-    *this = JSValue(static_cast<int32_t>(i));
-}
-
-inline JSValue::JSValue(unsigned long i)
-{
-    if (static_cast<uint32_t>(i) != i) {
-        *this = JSValue(EncodeAsDouble, static_cast<double>(i));
-        return;
-    }
-    *this = JSValue(static_cast<uint32_t>(i));
-}
-
-inline JSValue::JSValue(long long i)
-{
-    if (static_cast<int32_t>(i) != i) {
-        *this = JSValue(EncodeAsDouble, static_cast<double>(i));
-        return;
-    }
-    *this = JSValue(static_cast<int32_t>(i));
-}
-
-inline JSValue::JSValue(unsigned long long i)
-{
-    if (static_cast<uint32_t>(i) != i) {
-        *this = JSValue(EncodeAsDouble, static_cast<double>(i));
-        return;
-    }
-    *this = JSValue(static_cast<uint32_t>(i));
-}
-
-inline JSValue::JSValue(double d)
-{
-    if (canBeStrictInt32(d)) {
-        *this = JSValue(static_cast<int32_t>(d));
-        return;
-    }
-    *this = JSValue(EncodeAsDouble, d);
-}
-
-inline EncodedJSValue JSValue::encode(JSValue value)
-{
-    return value.u.asInt64;
-}
-
-inline JSValue JSValue::decode(EncodedJSValue encodedJSValue)
-{
-    JSValue v;
-    v.u.asInt64 = encodedJSValue;
-    return v;
-}
-
 #if USE(JSVALUE32_64)
-inline JSValue::JSValue()
-{
-    u.asBits.tag = EmptyValueTag;
-    u.asBits.payload = 0;
-}
-
-inline JSValue::JSValue(JSNullTag)
-{
-    u.asBits.tag = NullTag;
-    u.asBits.payload = 0;
-}
-    
-inline JSValue::JSValue(JSUndefinedTag)
-{
-    u.asBits.tag = UndefinedTag;
-    u.asBits.payload = 0;
-}
-    
-inline JSValue::JSValue(JSTrueTag)
-{
-    u.asBits.tag = BooleanTag;
-    u.asBits.payload = 1;
-}
-    
-inline JSValue::JSValue(JSFalseTag)
-{
-    u.asBits.tag = BooleanTag;
-    u.asBits.payload = 0;
-}
-
-inline JSValue::JSValue(HashTableDeletedValueTag)
-{
-    u.asBits.tag = DeletedValueTag;
-    u.asBits.payload = 0;
-}
-
-inline JSValue::JSValue(JSCell* ptr)
-{
-    if (ptr)
-        u.asBits.tag = CellTag;
-    else
-        u.asBits.tag = EmptyValueTag;
-    u.asBits.payload = reinterpret_cast<int32_t>(ptr);
-}
-
-inline JSValue::JSValue(const JSCell* ptr)
-{
-    if (ptr)
-        u.asBits.tag = CellTag;
-    else
-        u.asBits.tag = EmptyValueTag;
-    u.asBits.payload = reinterpret_cast<int32_t>(const_cast<JSCell*>(ptr));
-}
-
-inline JSValue::operator bool() const
-{
-    ASSERT(tag() != DeletedValueTag);
-    return tag() != EmptyValueTag;
-}
-
-inline bool JSValue::operator==(const JSValue& other) const
-{
-    return u.asInt64 == other.u.asInt64;
-}
-
-inline bool JSValue::isEmpty() const
-{
-    return tag() == EmptyValueTag;
-}
-
-inline bool JSValue::isUndefined() const
-{
-    return tag() == UndefinedTag;
-}
-
-inline bool JSValue::isNull() const
-{
-    return tag() == NullTag;
-}
-
-inline bool JSValue::isUndefinedOrNull() const
-{
-    return isUndefined() || isNull();
-}
-
-inline bool JSValue::isCell() const
-{
-    return tag() == CellTag;
-}
-
-inline bool JSValue::isInt32() const
-{
-    return tag() == Int32Tag;
-}
-
-inline bool JSValue::isDouble() const
-{
-    return tag() < LowestTag;
-}
-
-inline bool JSValue::isTrue() const
-{
-    return tag() == BooleanTag && payload();
-}
-
-inline bool JSValue::isFalse() const
-{
-    return tag() == BooleanTag && !payload();
-}
-
-inline uint32_t JSValue::tag() const
-{
-    return u.asBits.tag;
-}
-    
-inline int32_t JSValue::payload() const
-{
-    return u.asBits.payload;
-}
-    
-inline int32_t JSValue::asInt32() const
-{
-    ASSERT(isInt32());
-    return u.asBits.payload;
-}
-    
-inline double JSValue::asDouble() const
-{
-    ASSERT(isDouble());
-    return u.asDouble;
-}
-    
-ALWAYS_INLINE JSCell* JSValue::asCell() const
-{
-    ASSERT(isCell());
-    return reinterpret_cast<JSCell*>(u.asBits.payload);
-}
-
 ALWAYS_INLINE JSBigInt* JSValue::asHeapBigInt() const
 {
     ASSERT(isHeapBigInt());
     return reinterpret_cast<JSBigInt*>(u.asBits.payload);
 }
-
-ALWAYS_INLINE JSValue::JSValue(EncodeAsDoubleTag, double d)
-{
-    ASSERT(!isImpureNaN(d));
-    u.asDouble = d;
-}
-
-inline JSValue::JSValue(int i)
-{
-    u.asBits.tag = Int32Tag;
-    u.asBits.payload = i;
-}
-
-inline JSValue::JSValue(int32_t tag, int32_t payload)
-{
-    u.asBits.tag = tag;
-    u.asBits.payload = payload;
-}
-
-inline bool JSValue::isNumber() const
-{
-    return isInt32() || isDouble();
-}
-
-inline bool JSValue::isBoolean() const
-{
-    return tag() == BooleanTag;
-}
-
-inline bool JSValue::asBoolean() const
-{
-    ASSERT(isBoolean());
-    return payload();
-}
-
 #else // !USE(JSVALUE32_64) i.e. USE(JSVALUE64)
-
-// 0x0 can never occur naturally because it has a tag of 00, indicating a pointer value, but a payload of 0x0, which is in the (invalid) zero page.
-inline JSValue::JSValue()
-{
-    u.asInt64 = ValueEmpty;
-}
-
-// 0x4 can never occur naturally because it has a tag of 00, indicating a pointer value, but a payload of 0x4, which is in the (invalid) zero page.
-inline JSValue::JSValue(HashTableDeletedValueTag)
-{
-    u.asInt64 = ValueDeleted;
-}
-
-inline JSValue::JSValue(JSCell* ptr)
-{
-    u.asInt64 = reinterpret_cast<uintptr_t>(ptr);
-}
-
-inline JSValue::JSValue(const JSCell* ptr)
-{
-    u.asInt64 = reinterpret_cast<uintptr_t>(const_cast<JSCell*>(ptr));
-}
-
-inline JSValue::operator bool() const
-{
-    return u.asInt64;
-}
-
-inline bool JSValue::operator==(const JSValue& other) const
-{
-    return u.asInt64 == other.u.asInt64;
-}
-
-inline bool JSValue::isEmpty() const
-{
-    return u.asInt64 == ValueEmpty;
-}
-
-inline bool JSValue::isUndefined() const
-{
-    return asValue() == JSValue(JSUndefined);
-}
-
-inline bool JSValue::isNull() const
-{
-    return asValue() == JSValue(JSNull);
-}
-
-inline bool JSValue::isTrue() const
-{
-    return asValue() == JSValue(JSTrue);
-}
-
-inline bool JSValue::isFalse() const
-{
-    return asValue() == JSValue(JSFalse);
-}
-
-inline bool JSValue::asBoolean() const
-{
-    ASSERT(isBoolean());
-    return asValue() == JSValue(JSTrue);
-}
-
-inline int32_t JSValue::asInt32() const
-{
-    ASSERT(isInt32());
-    return static_cast<int32_t>(u.asInt64);
-}
-
-inline bool JSValue::isDouble() const
-{
-    return isNumber() && !isInt32();
-}
-
-inline JSValue::JSValue(JSNullTag)
-{
-    u.asInt64 = ValueNull;
-}
-    
-inline JSValue::JSValue(JSUndefinedTag)
-{
-    u.asInt64 = ValueUndefined;
-}
-
-inline JSValue::JSValue(JSTrueTag)
-{
-    u.asInt64 = ValueTrue;
-}
-
-inline JSValue::JSValue(JSFalseTag)
-{
-    u.asInt64 = ValueFalse;
-}
-
-inline bool JSValue::isUndefinedOrNull() const
-{
-    // Undefined and null share the same value, bar the 'undefined' bit in the extended tag.
-    return (u.asInt64 & ~UndefinedTag) == ValueNull;
-}
-
-inline bool JSValue::isBoolean() const
-{
-    return (u.asInt64 & ~1) == ValueFalse;
-}
-
-inline bool JSValue::isCell() const
-{
-    return !(u.asInt64 & NotCellMask);
-}
-
-inline bool JSValue::isInt32() const
-{
-    return (u.asInt64 & NumberTag) == NumberTag;
-}
-
-inline int64_t reinterpretDoubleToInt64(double value)
-{
-    return std::bit_cast<int64_t>(value);
-}
-inline double reinterpretInt64ToDouble(int64_t value)
-{
-    return std::bit_cast<double>(value);
-}
-
-ALWAYS_INLINE JSValue::JSValue(EncodeAsDoubleTag, double d)
-{
-    ASSERT(!isImpureNaN(d));
-    u.asInt64 = reinterpretDoubleToInt64(d) + JSValue::DoubleEncodeOffset;
-}
-
-inline JSValue::JSValue(int i)
-{
-    u.asInt64 = JSValue::NumberTag | static_cast<uint32_t>(i);
-}
-
-inline double JSValue::asDouble() const
-{
-    ASSERT(isDouble());
-    return reinterpretInt64ToDouble(u.asInt64 - JSValue::DoubleEncodeOffset);
-}
-
-inline bool JSValue::isNumber() const
-{
-    return u.asInt64 & JSValue::NumberTag;
-}
-
-ALWAYS_INLINE JSCell* JSValue::asCell() const
-{
-    ASSERT(isCell());
-    return u.ptr;
-}
-
 ALWAYS_INLINE JSBigInt* JSValue::asHeapBigInt() const
 {
     ASSERT(isHeapBigInt());
     return static_cast<JSBigInt*>(u.ptr);
 }
-
 #endif // USE(JSVALUE64)
 
-#if USE(BIGINT32)
-inline JSValue::JSValue(EncodeAsBigInt32Tag, int32_t value)
+// ECMA 11.9.3
+inline bool JSValue::equal(JSGlobalObject* globalObject, JSValue v1, JSValue v2)
 {
-    uint64_t shiftedValue = static_cast<uint64_t>(static_cast<uint32_t>(value)) << 16;
-    ASSERT(!(shiftedValue & NumberTag));
-    u.asInt64 = shiftedValue | BigInt32Tag;
-}
-#endif // USE(BIGINT32)
+    if (v1.isInt32() && v2.isInt32())
+        return v1 == v2;
 
-#if ENABLE(WEBASSEMBLY) && USE(JSVALUE32_64)
-inline JSValue::JSValue(EncodeAsUnboxedFloatTag, float value)
-{
-    u.asBits.payload = std::bit_cast<int32_t>(value);
+    return equalSlowCase(globalObject, v1, v2);
 }
-#endif
-
-inline int64_t tryConvertToInt52(double number)
-{
-    if (number != number)
-        return JSValue::notInt52;
-#if OS(WINDOWS) && CPU(X86)
-    // The VS Compiler for 32-bit builds generates a floating point error when attempting to cast
-    // from an infinity to a 64-bit integer. We leave this routine with the floating point error
-    // left in a register, causing undefined behavior in later floating point operations.
-    //
-    // To avoid this issue, we check for infinity here, and return false in that case.
-    if (std::isinf(number))
-        return JSValue::notInt52;
-#endif
-    int64_t asInt64 = static_cast<int64_t>(number);
-    if (asInt64 != number)
-        return JSValue::notInt52;
-    if (!asInt64 && std::signbit(number))
-        return JSValue::notInt52;
-    if (asInt64 >= (static_cast<int64_t>(1) << (JSValue::numberOfInt52Bits - 1)))
-        return JSValue::notInt52;
-    if (asInt64 < -(static_cast<int64_t>(1) << (JSValue::numberOfInt52Bits - 1)))
-        return JSValue::notInt52;
-    return asInt64;
-}
-
-inline bool isInt52(double number)
-{
-    return tryConvertToInt52(number) != JSValue::notInt52;
-}
-
-inline bool JSValue::isAnyInt() const
-{
-    if (isInt32())
-        return true;
-    if (!isNumber())
-        return false;
-    return isInt52(asDouble());
-}
-
-inline int64_t JSValue::asAnyInt() const
-{
-    ASSERT(isAnyInt());
-    if (isInt32())
-        return asInt32();
-    return static_cast<int64_t>(asDouble());
-}
-
-inline bool JSValue::isInt32AsAnyInt() const
-{
-    if (!isAnyInt())
-        return false;
-    int64_t value = asAnyInt();
-    return value >= INT32_MIN && value <= INT32_MAX;
-}
-
-inline int32_t JSValue::asInt32AsAnyInt() const
-{
-    ASSERT(isInt32AsAnyInt());
-    if (isInt32())
-        return asInt32();
-    return static_cast<int32_t>(asDouble());
-}
-
-inline bool JSValue::isUInt32AsAnyInt() const
-{
-    if (!isAnyInt())
-        return false;
-    int64_t value = asAnyInt();
-    return value >= 0 && value <= UINT32_MAX;
-}
-
-inline uint32_t JSValue::asUInt32AsAnyInt() const
-{
-    ASSERT(isUInt32AsAnyInt());
-    if (isUInt32())
-        return asUInt32();
-    return static_cast<uint32_t>(asDouble());
-}
-
-inline bool JSValue::isString() const
-{
-    return isCell() && asCell()->isString();
-}
-
-inline bool JSValue::isBigInt() const
-{
-    return isBigInt32() || isHeapBigInt();
-}
-
-inline bool JSValue::isHeapBigInt() const
-{
-    return isCell() && asCell()->isHeapBigInt();
-}
-
-inline bool JSValue::isBigInt32() const
-{
-#if USE(BIGINT32)
-    return (u.asInt64 & BigInt32Mask) == BigInt32Tag;
-#else
-    return false;
-#endif
-}
-
-#if USE(BIGINT32)
-inline int32_t JSValue::bigInt32AsInt32() const
-{
-    ASSERT(isBigInt32());
-    return static_cast<int32_t>(u.asInt64 >> 16);
-}
-#endif // USE(BIGINT32)
 
 inline bool JSValue::isZeroBigInt() const
 {
@@ -738,49 +190,9 @@ inline bool JSValue::isNegativeBigInt() const
     return asHeapBigInt()->sign();
 }
 
-inline bool JSValue::isSymbol() const
-{
-    return isCell() && asCell()->isSymbol();
-}
-
-inline bool JSValue::isPrimitive() const
-{
-    return !isCell() || asCell()->isString() || asCell()->isSymbol() || asCell()->isHeapBigInt();
-}
-
-inline bool JSValue::isGetterSetter() const
-{
-    return isCell() && asCell()->isGetterSetter();
-}
-
-inline bool JSValue::isCustomGetterSetter() const
-{
-    return isCell() && asCell()->isCustomGetterSetter();
-}
-
-inline bool JSValue::isObject() const
-{
-    return isCell() && asCell()->isObject();
-}
-
-inline bool JSValue::getString(JSGlobalObject* globalObject, String& s) const
-{
-    return isCell() && asCell()->getString(globalObject, s);
-}
-
-inline String JSValue::getString(JSGlobalObject* globalObject) const
-{
-    return isCell() ? asCell()->getString(globalObject) : String();
-}
-
 template <typename Base> String HandleConverter<Base, Unknown>::getString(JSGlobalObject* globalObject) const
 {
     return jsValue().getString(globalObject);
-}
-
-inline JSObject* JSValue::getObject() const
-{
-    return isCell() ? asCell()->getObject() : nullptr;
 }
 
 ALWAYS_INLINE bool JSValue::getUInt32(uint32_t& v) const
@@ -832,11 +244,6 @@ ALWAYS_INLINE JSValue JSValue::toPropertyKeyValue(JSGlobalObject* globalObject) 
     RELEASE_AND_RETURN(scope, primitive.toString(globalObject));
 }
 
-inline JSValue JSValue::toPrimitive(JSGlobalObject* globalObject, PreferredPrimitiveType preferredType) const
-{
-    return isCell() ? asCell()->toPrimitive(globalObject, preferredType) : asValue();
-}
-
 inline PreferredPrimitiveType toPreferredPrimitiveType(JSGlobalObject* globalObject, JSValue value)
 {
     VM& vm = getVM(globalObject);
@@ -859,15 +266,6 @@ inline PreferredPrimitiveType toPreferredPrimitiveType(JSGlobalObject* globalObj
 
     throwTypeError(globalObject, scope, "Expected primitive hint to match one of 'default', 'number', 'string'."_s);
     return NoPreference;
-}
-
-ALWAYS_INLINE double JSValue::toNumber(JSGlobalObject* globalObject) const
-{
-    if (isInt32())
-        return asInt32();
-    if (isDouble())
-        return asDouble();
-    return toNumberSlowCase(globalObject);
 }
 
 ALWAYS_INLINE JSValue JSValue::toNumeric(JSGlobalObject* globalObject) const
@@ -925,20 +323,6 @@ ALWAYS_INLINE JSValue JSValue::toBigIntOrInt32(JSGlobalObject* globalObject) con
     return jsNumber(value);
 }
 
-inline bool JSValue::toBoolean(JSGlobalObject* globalObject) const
-{
-    if (isInt32())
-        return asInt32();
-    if (isDouble())
-        return asDouble() > 0.0 || asDouble() < 0.0; // false for NaN
-    if (isCell())
-        return asCell()->toBoolean(globalObject);
-#if USE(BIGINT32)
-    if (isBigInt32())
-        return !!bigInt32AsInt32();
-#endif
-    return isTrue(); // false, null, and undefined all convert to false.
-}
 
 inline JSString* JSValue::toString(JSGlobalObject* globalObject) const
 {
@@ -963,53 +347,6 @@ inline String JSValue::toWTFString(JSGlobalObject* globalObject) const
     return toWTFStringSlowCase(globalObject);
 }
 
-inline JSObject* JSValue::toObject(JSGlobalObject* globalObject) const
-{
-    return isCell() ? asCell()->toObject(globalObject) : toObjectSlowCase(globalObject);
-}
-
-inline bool JSValue::isCallable() const
-{
-    return isCell() && asCell()->isCallable();
-}
-
-template<Concurrency concurrency>
-inline TriState JSValue::isCallableWithConcurrency() const
-{
-    if (!isCell())
-        return TriState::False;
-    return asCell()->isCallableWithConcurrency<concurrency>();
-}
-
-inline bool JSValue::isConstructor() const
-{
-    return isCell() && asCell()->isConstructor();
-}
-
-template<Concurrency concurrency>
-inline TriState JSValue::isConstructorWithConcurrency() const
-{
-    if (!isCell())
-        return TriState::False;
-    return asCell()->isConstructorWithConcurrency<concurrency>();
-}
-
-// this method is here to be after the inline declaration of JSCell::inherits
-inline bool JSValue::inherits(const ClassInfo* classInfo) const
-{
-    return isCell() && asCell()->inherits(classInfo);
-}
-
-template<typename Target>
-inline bool JSValue::inherits() const
-{
-    return isCell() && asCell()->inherits<Target>();
-}
-
-inline const ClassInfo* JSValue::classInfoOrNull() const
-{
-    return isCell() ? asCell()->classInfo() : nullptr;
-}
 
 inline JSValue JSValue::toThis(JSGlobalObject* globalObject, ECMAMode ecmaMode) const
 {
@@ -1182,22 +519,6 @@ ALWAYS_INLINE JSValue JSValue::getPrototype(JSGlobalObject* globalObject) const
     return synthesizePrototype(globalObject);
 }
 
-inline Structure* JSValue::structureOrNull() const
-{
-    if (isCell())
-        return asCell()->structure();
-    return nullptr;
-}
-
-// ECMA 11.9.3
-inline bool JSValue::equal(JSGlobalObject* globalObject, JSValue v1, JSValue v2)
-{
-    if (v1.isInt32() && v2.isInt32())
-        return v1 == v2;
-
-    return equalSlowCase(globalObject, v1, v2);
-}
-
 ALWAYS_INLINE bool JSValue::equalSlowCaseInline(JSGlobalObject* globalObject, JSValue v1, JSValue v2)
 {
     VM& vm = getVM(globalObject);
@@ -1355,13 +676,6 @@ inline bool JSValue::strictEqual(JSGlobalObject* globalObject, JSValue v1, JSVal
         return strictEqualForCells(globalObject, v1.asCell(), v2.asCell());
 
     return v1 == v2;
-}
-
-inline int32_t JSValue::asInt32ForArithmetic() const
-{
-    if (isBoolean())
-        return asBoolean();
-    return asInt32();
 }
 
 inline TriState JSValue::pureStrictEqual(JSValue v1, JSValue v2)
