+++ Source/JavaScriptCore/runtime/WriteBarrier.h
@@ -179,11 +179,11 @@ public:
     void setStartingValue(JSValue value) { m_value = JSValue::encode(value); }
     bool isNumber() const { return get().isNumber(); }
     bool isInt32() const { return get().isInt32(); }
-    bool isObject() const { return get().isObject(); }
-    bool isNull() const { return get().isNull(); }
-    bool isGetterSetter() const { return get().isGetterSetter(); }
-    bool isCustomGetterSetter() const { return get().isCustomGetterSetter(); }
-    
+    inline bool isObject() const; // Defined in WriteBarrierInlines.h
+    inline bool isNull() const; // Defined in WriteBarrierInlines.h
+    inline bool isGetterSetter() const; // Defined in WriteBarrierInlines.h
+    inline bool isCustomGetterSetter() const; // Defined in WriteBarrierInlines.h
+
     JSValue* slot() const
     { 
         return std::bit_cast<JSValue*>(&m_value);
